it has a problem with the conversion, I converted string to long
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/LHCbDIRAC/trunk@9478 4525493e-7705-40b1-a816-d608a930855b

+++ LHCbDIRAC/BookkeepingSystem/Client/BookkeepingClient.py
@@ -54,7 +54,7 @@ class BookkeepingClient:
   
   #############################################################################
   def getSpecificFiles(self, configName, configVersion, programName, programVersion, fileType, eventTypeId, production):
-    result = self.server.getSpecificFiles(configName, configVersion, programName, programVersion, fileType, eventTypeId, production)
+    result = self.server.getSpecificFiles(configName, configVersion, programName, programVersion, fileType, long(eventTypeId), long(production))
     return result
   
   #############################################################################
